+++ src/CapabilityService.Tests/TestDoubles/StubStartUp.cs
@@ -1,5 +1,7 @@
 ﻿using DFDS.CapabilityService.Tests.Infrastructure.Authentication;
 using DFDS.CapabilityService.WebApi;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc.Authorization;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 
@@ -13,12 +15,22 @@ namespace DFDS.CapabilityService.Tests.TestDoubles
 
 		protected override void ConfigureAuth(IServiceCollection services)
 		{
+			services.AddMvc(config =>
+			{
+				var policy = new AuthorizationPolicyBuilder()
+					.AddAuthenticationSchemes(MockAuthenticationSchemeOptions.SchemeName)
+					.RequireAuthenticatedUser()
+					.Build();
+
+				config.Filters.Add(new AuthorizeFilter(policy));
+			});
+
 			services.AddAuthentication(options =>
 			{
-				options.DefaultAuthenticateScheme = LocalAuthenticationSchemeOptions.SchemeName;
-				options.DefaultChallengeScheme = LocalAuthenticationSchemeOptions.SchemeName;
+				options.DefaultAuthenticateScheme = MockAuthenticationSchemeOptions.SchemeName;
+				options.DefaultChallengeScheme = MockAuthenticationSchemeOptions.SchemeName;
 			})
-			.AddScheme<LocalAuthenticationSchemeOptions, LocalAuthenticationHandler>(LocalAuthenticationSchemeOptions.SchemeName, null);
+			.AddScheme<MockAuthenticationSchemeOptions, MockAuthenticationHandler>(MockAuthenticationSchemeOptions.SchemeName, null);
 		}
 	}
 }
